#45 launcher working correctly
PDash can load projects from files,
When the user scans a project a PDashScan.bin gets created next to the project containing all the tree with the paths and all the info.
Next time, when the user opens PDash, an init screen will have the projects history where the user can doble click on the project, even if the project is not opened, PDash will open the project and load everything from the scanned file.