[FIX] Fixed Scripting for creating constants, if the user added clusters or nested controls the scripting failed, added extra validations and it is working
also highlight the new constant and error after Panther Dashboard create the new vi.
Renamed the menu from remove module to exclude module by Matthias Baudot sugestion, it makes the user to think that panther dashboard will delete the module from the project and that is not the case, this functionality is to remove/exclude the module from the scan.